Rename and move PipelineProcessConfigurationUtils
